phyzioline/requirements.txt
# Django Core
Django==5.2.8

# Django REST Framework
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

# CORS Headers (للسماح بالاتصال من Frontend)
django-cors-headers==4.9.0

# Filtering
django-filter==24.3

# Database (PostgreSQL for production)
psycopg2-binary==2.9.11

# Image Processing
Pillow>=10.0.0

# Utilities
requests==2.32.3

# Environment variables (اختياري - للإنتاج)
# python-decouple==3.8

# Additional packages (للمستقبل)
# celery==5.3.4
# redis==5.0.1
# channels==4.0.0
# channels-redis==4.2.0

